refactor: 리뷰 정보를 식품 도메인에 물어보기
