+++ backend/src/test/resources/truncate.sql
@@ -1,4 +1,6 @@
 SET FOREIGN_KEY_CHECKS = 0;
+truncate table review;
+truncate table member;
 truncate table pet_food;
 truncate table keyword;
 truncate table brand;
